requirements.txt
torch
transformers
numpy
datasets
jinja2
faiss-cpu
requests
flask
sentence_transformers
fastapi
uvicorn
httpx
